chore: move mocks into src folder
