+++ src/actions/__mocks__/getPhotosDetails.revolve.ts
