+++ test/actions/lists.test.ts
@@ -4,7 +4,7 @@ import {
   mockingFailingNotValidFetch,
   mockingFetch,
 } from '../../src/utils/test'
-import { payload as payloadGetLists } from './__mocks__/getLists.resolve'
+import { payload as payloadGetLists } from '../../src/actions/__mocks__/getLists.resolve'
 
 describe('actions/lists', () => {
   test.each`
